requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
decorator==4.4.2
Django==4.1.7
django-celery-beat==2.4.0
django-cleanup==7.0.0
django-timezone-field==5.0
gunicorn==20.1.0
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
kombu==5.2.4
moviepy==1.0.3
numpy==1.24.2
pafy==0.5.5
Pillow==9.4.0
proglog==0.1.10
prompt-toolkit==3.0.38
psycopg2==2.9.6
pydub==0.25.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytube==12.1.3
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
SpeechRecognition==3.9.0
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
youtube-dl==2020.12.31